feat: adjust home page content
