+++ src/lib/constants/routes/public.ts
@@ -1,1 +1,1 @@
-export const PUBLIC_ROUTES = ['/auth', '/intro', '/privacypolicy'];
+export const PUBLIC_ROUTES = ['/auth', '/home', '/privacypolicy'];
