+++ src/lib/styles/theme/index.ts
@@ -4,12 +4,14 @@ import { colors } from './colors';
 import { components } from './components';
 import { config } from './config';
 import { fonts } from './fonts';
+import { styles } from './styles';
 
 const customTheme = extendTheme({
   fonts,
   components,
   config,
   colors,
+  styles,
 });
 
 export default customTheme;
