+++ src/lib/styles/theme/styles.ts
@@ -1,0 +1,9 @@
+import type { DeepPartial, Theme } from '@chakra-ui/react';
+
+export const styles: DeepPartial<Theme['styles']> = {
+  global: {
+    body: {
+      overflowY: 'scroll',
+    },
+  },
+};
