+++ src/pages/home.ts
@@ -1,0 +1,3 @@
+import Intro from '~/lib/pages/home';
+
+export default Intro;
