--- src/pages/intro.ts
@@ -1,3 +1,0 @@
-import Intro from '~/lib/pages/intro';
-
-export default Intro;
